fix alpha blending and normal mapping

+++ src/resources/cubemap.hpp
@@ -224,11 +224,11 @@ private:
 			};
 			for (const auto& view_matrix : view_matrices) {
 				glUniformMatrix3fv(this->view_matrix.location(), 1, GL_FALSE, glm::value_ptr(view_matrix));
-				glFramebufferTexture2D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, target, result.get(), level);
+				glFramebufferTexture2D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, target, result.get(), static_cast<GLint>(level));
 				opengl_context::check_status();
 				opengl_context::check_framebuffer_status();
 				glDrawArrays(cubemap_mesh::primitive_type, 0, static_cast<GLsizei>(cubemap_mesh::vertices.size()));
-				glFramebufferTexture2D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, target, 0, level);
+				glFramebufferTexture2D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, target, 0, static_cast<GLint>(level));
 				++target;
 			}
 		}
